+++ tests/Integration/ExceptionsTest.php
@@ -25,11 +25,12 @@ use Illuminate\Auth\AuthenticationException;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Foundation\Http\Exceptions\MaintenanceModeException;
 use Illuminate\Session\TokenMismatchException;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Route;
 use Illuminate\Support\MessageBag;
 use Illuminate\Validation\ValidationException;
 use LaravelJsonApi\Core\Exceptions\JsonApiException;
-use LaravelJsonApi\Spec\UnexpectedDocumentException;
 use Symfony\Component\HttpFoundation\Exception\BadRequestException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\HttpKernel\Exception\MethodNotAllowedHttpException;
@@ -472,104 +473,10 @@ class ExceptionsTest extends TestCase
             ->assertExactJson($expected);
     }
 
-    public function testUnexpectedDocument(): void
+    public function testDefaultExceptionWithoutDebug(): void
     {
-        $this->ex = new UnexpectedDocumentException('That document is wrong.');
+        config()->set('app.debug', false);
 
-        $expected = [
-            'errors' => [
-                [
-                    'detail' => 'That document is wrong.',
-                    'title' => 'Invalid JSON',
-                    'status' => '400',
-                ],
-            ],
-            'jsonapi' => [
-                'version' => '1.0',
-            ],
-        ];
-
-        $this->get('/test', ['Accept' => 'application/vnd.api+json'])
-            ->assertStatus(400)
-            ->assertHeader('Content-Type', 'application/vnd.api+json')
-            ->assertExactJson($expected);
-    }
-
-    public function testUnexpectedDocumentWithoutMessage(): void
-    {
-        $this->ex = new UnexpectedDocumentException();
-
-        $expected = [
-            'errors' => [
-                [
-                    'title' => 'Invalid JSON',
-                    'status' => '400',
-                ],
-            ],
-            'jsonapi' => [
-                'version' => '1.0',
-            ],
-        ];
-
-        $this->get('/test', ['Accept' => 'application/vnd.api+json'])
-            ->assertStatus(400)
-            ->assertHeader('Content-Type', 'application/vnd.api+json')
-            ->assertExactJson($expected);
-    }
-
-    public function testUnexpectedDocumentWithJsonException(): void
-    {
-        $json = new \JsonException('Bad format.', 4);
-
-        $this->ex = new UnexpectedDocumentException('That document is wrong.', 0, $json);
-
-        $expected = [
-            'errors' => [
-                [
-                    'code' => '4',
-                    'detail' => 'Bad format.',
-                    'title' => 'Invalid JSON',
-                    'status' => '400',
-                ],
-            ],
-            'jsonapi' => [
-                'version' => '1.0',
-            ],
-        ];
-
-        $this->get('/test', ['Accept' => 'application/vnd.api+json'])
-            ->assertStatus(400)
-            ->assertHeader('Content-Type', 'application/vnd.api+json')
-            ->assertExactJson($expected);
-    }
-
-    public function testUnexpectedDocumentWithJsonExceptionWithoutMessage(): void
-    {
-        $json = new \JsonException('', 4);
-
-        $this->ex = new UnexpectedDocumentException('That document is wrong.', 0, $json);
-
-        $expected = [
-            'errors' => [
-                [
-                    'code' => '4',
-                    'title' => 'Invalid JSON',
-                    'status' => '400',
-                ],
-            ],
-            'jsonapi' => [
-                'version' => '1.0',
-            ],
-        ];
-
-        $this->get('/test', ['Accept' => 'application/vnd.api+json'])
-            ->assertStatus(400)
-            ->assertHeader('Content-Type', 'application/vnd.api+json')
-            ->assertExactJson($expected);
-    }
-
-    public function testGenericException(): void
-    {
         $this->ex = new \Exception('Boom.');
 
         $expected = [
@@ -577,6 +484,40 @@ class ExceptionsTest extends TestCase
                 [
                     'title' => 'Internal Server Error',
                     'status' => '500',
+                ],
+            ],
+            'jsonapi' => [
+                'version' => '1.0',
+            ],
+        ];
+
+        $this->get('/test', ['Accept' => 'application/vnd.api+json'])
+            ->assertStatus(500)
+            ->assertHeader('Content-Type', 'application/vnd.api+json')
+            ->assertExactJson($expected);
+    }
+
+    public function testDefaultExceptionWithDebug(): void
+    {
+        config()->set('app.debug', true);
+
+        $this->ex = $ex = new \LogicException('Boom.', 99);
+
+        $expected = [
+            'errors' => [
+                [
+                    'code' => (string) $ex->getCode(),
+                    'detail' => $ex->getMessage(),
+                    'meta' => [
+                        'exception' => get_class($ex),
+                        'file' => $ex->getFile(),
+                        'line' => $ex->getLine(),
+                        'trace' => Collection::make($ex->getTrace())
+                            ->map(fn($trace) => Arr::except($trace, ['args']))
+                            ->all(),
+                    ],
+                    'status' => '500',
+                    'title' => 'Internal Server Error',
                 ],
             ],
             'jsonapi' => [