+++ CHANGELOG.md
@@ -12,6 +12,12 @@ All notable changes to this project will be documented in this file. This projec
   the actual exception message to the `detail` member. The `meta` member will contain the exception class, file, line
   number and stack trace.
 
+### Removed
+
+- This package no longer handles the `UnexpectedDocumentException` from the `laravel-json-api/spec` package. The
+  exception has been removed from that package, and it instead throws a `JsonApiException` if it cannot decode a JSON
+  string.
+
 ## [1.0.0-beta.3] - 2021-04-26
 
 ### Added